backend/requirements.txt
requests>=2.31.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
selenium>=4.15.0
webdriver-manager>=4.0.0 